requirements.txt
numpy==1.25.2
Pillow==9.4.0
torch==2.1.0
torchinfo==1.8.0
torchvision==0.16.0
piq==0.8.0
